Moved HasWireNumber into common module
