+++ kbl/kbl-v25/src/main/resources/kbl25/KBL25_SR1-ext.xjb
@@ -29,6 +29,10 @@
         <inheritance:implements>com.foursoft.harness.kbl.common.HasIdentification</inheritance:implements>
     </jxb:bindings>
 
+    <jxb:bindings multiple="true" node="//xs:complexType[.//xs:element[@name='Wire_number' ]]">
+        <inheritance:implements>com.foursoft.harness.kbl.v25.HasWireNumber</inheritance:implements>
+    </jxb:bindings>
+
     <jxb:bindings multiple="true"
             node="//xs:complexType[@name='Accessory_occurrence'
                     or @name='Assembly_part_occurrence'
@@ -122,8 +126,4 @@
         <inheritance:implements>com.foursoft.harness.kbl.v25.HasRelatedOccurrence</inheritance:implements>
     </jxb:bindings>
 
-    <jxb:bindings multiple="true" node="//xs:complexType[.//xs:element[@name='Wire_number' ]]">
-        <inheritance:implements>com.foursoft.harness.kbl.v25.HasWireNumber</inheritance:implements>
-    </jxb:bindings>
-    
 </jxb:bindings>
